#89 Decompose book form
