+++ extension/ts/entities/bookForm/types.ts
@@ -1,0 +1,4 @@
+type FormAreaElement = HTMLTextAreaElement | HTMLInputElement | HTMLSelectElement;
+type FormData = Record<string, Record<string, any>>;
+
+export type {FormAreaElement, FormData};
