+++ extension/ts/entities/bookForm/ui.ts
@@ -1,0 +1,15 @@
+const show = (element: Element) => {
+	// LibraryThing collections checkboxes are sometimes hidden beneath a div that is not visible
+	// THIS IS BRITTLE and relies on the specific markup tree of LibraryThing
+	const hiddenAncestor = element.closest<HTMLElement>('div[style="display:none;"]');
+	if (hiddenAncestor) {
+		hiddenAncestor.style.display = "";
+	}
+};
+
+const ensureAuthorInputCount = (count: number): void => {
+	// TODO
+	console.log(count);
+};
+
+export {show, ensureAuthorInputCount};
